+++ plugins/modules/harbor_config.py
@@ -1,5 +1,5 @@
 import copy
-from plugins.modules.harbor_base import HarborBaseModule
+from plugins.module_utils.harbor_base import HarborBaseModule
 import json
 from ansible.module_utils.basic import AnsibleModule
