+++ plugins/modules/harbor_garbage_collection.py
@@ -10,7 +10,7 @@ import json
 
 
 from ansible.module_utils.basic import AnsibleModule
-from plugins.modules.harbor_base import \
+from plugins.module_utils.harbor_base import \
     HarborBaseModule
 
 DOCUMENTATION = '''
