+++ plugins/modules/harbor_project.py
@@ -8,7 +8,7 @@
 import copy
 import json
 from ansible.module_utils.basic import AnsibleModule
-from plugins.modules.harbor_base import HarborBaseModule
+from plugins.module_utils.harbor_base import HarborBaseModule
 
 DOCUMENTATION = '''
 ---
