+++ plugins/modules/harbor_project_member.py
@@ -6,7 +6,7 @@
 # https://www.gnu.org/licenses/gpl-3.0.txt)
 import copy
 from ansible.module_utils.basic import AnsibleModule
-from plugins.modules.harbor_base import \
+from plugins.module_utils.harbor_base import \
     HarborBaseModule
 
 DOCUMENTATION = '''
